V114.00	Fixed implicit cast warnings 	(on display_support.c,h display_if_basis.c,h font_if.c,h hw_config.c,h 	    pwr_support.c,h gnss_support.c,h uart_support_gps.c,h rtc_support.c,h 	    usb_msc_mass_mal.c,h usb_cdc_ecndp.c sdio_stm32f1.c,h syscalls.c 	    and many in "./lib/display"). 	Exchanged xprintf function for latest change(on xprintf.c,h). 	Fixed dropped IL Font(font_if_datatable.c,h). 	Fixed ".editorconfig" file for github code appearance.

+++ lib/STM32F10x_StdPeriph_Driver/src/misc.c
@@ -135,7 +135,7 @@ void NVIC_Init(NVIC_InitTypeDef* NVIC_InitStruct)
     tmppriority |=  NVIC_InitStruct->NVIC_IRQChannelSubPriority & tmpsub;
     tmppriority = tmppriority << 0x04;
         
-    NVIC->IPR[NVIC_InitStruct->NVIC_IRQChannel] = tmppriority; /* 20231218 Nemui CMSIS_V6 adopted. */
+    NVIC->IPR[NVIC_InitStruct->NVIC_IRQChannel] = (uint8_t)tmppriority; /* 20250618 Nemui CMSIS_V6 adopted. */
     
     /* Enable the Selected IRQ Channels --------------------------------------*/
     NVIC->ISER[NVIC_InitStruct->NVIC_IRQChannel >> 0x05] =
